APIharmo/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
tensorflow-cpu==2.20.0
numpy>=1.24.0
scikit-learn>=1.3.0
pydantic>=2.5.0